Fix #62: test reorganize
